Manual cherry of https://github.com/istio/istio/pull/52320
