+++ cni/pkg/nodeagent/informers_test.go
@@ -65,10 +65,7 @@ func TestExistingPodAddedWhenNsLabeled(t *testing.T) {
 		"",
 	).Return(nil)
 
-	server := &meshDataplane{
-		kubeClient: client.Kube(),
-		netServer:  fs,
-	}
+	server := getFakeDP(fs, client.Kube())
 
 	handlers := setupHandlers(ctx, client, server, "istio-system")
 	client.RunAndWait(ctx.Done())
@@ -127,10 +124,7 @@ func TestExistingPodAddedWhenDualStack(t *testing.T) {
 		"",
 	).Return(nil)
 
-	server := &meshDataplane{
-		kubeClient: client.Kube(),
-		netServer:  fs,
-	}
+	server := getFakeDP(fs, client.Kube())
 
 	fs.Start(ctx)
 	handlers := setupHandlers(ctx, client, server, "istio-system")
@@ -179,10 +173,7 @@ func TestExistingPodNotAddedIfNoIPInAnyStatusField(t *testing.T) {
 
 	fs := &fakeServer{}
 
-	server := &meshDataplane{
-		kubeClient: client.Kube(),
-		netServer:  fs,
-	}
+	server := getFakeDP(fs, client.Kube())
 
 	handlers := setupHandlers(ctx, client, server, "istio-system")
 	client.RunAndWait(ctx.Done())
@@ -244,10 +235,7 @@ func TestExistingPodRemovedWhenNsUnlabeled(t *testing.T) {
 		"",
 	).Return(nil)
 
-	server := &meshDataplane{
-		kubeClient: client.Kube(),
-		netServer:  fs,
-	}
+	server := getFakeDP(fs, client.Kube())
 
 	handlers := setupHandlers(ctx, client, server, "istio-system")
 	client.RunAndWait(ctx.Done())
@@ -337,10 +325,7 @@ func TestExistingPodRemovedWhenPodLabelRemoved(t *testing.T) {
 		"",
 	).Return(nil)
 
-	server := &meshDataplane{
-		kubeClient: client.Kube(),
-		netServer:  fs,
-	}
+	server := getFakeDP(fs, client.Kube())
 
 	handlers := setupHandlers(ctx, client, server, "istio-system")
 	client.RunAndWait(ctx.Done())
@@ -440,10 +425,7 @@ func TestJobPodRemovedWhenPodTerminates(t *testing.T) {
 		"",
 	).Return(nil)
 
-	server := &meshDataplane{
-		kubeClient: client.Kube(),
-		netServer:  fs,
-	}
+	server := getFakeDP(fs, client.Kube())
 
 	handlers := setupHandlers(ctx, client, server, "istio-system")
 	client.RunAndWait(ctx.Done())
@@ -556,10 +538,8 @@ func TestGetActiveAmbientPodSnapshotOnlyReturnsActivePods(t *testing.T) {
 	client := kube.NewFakeClient(ns, enrolledNotRedirected, redirectedNotEnrolled)
 	fs := &fakeServer{}
 	fs.Start(ctx)
-	server := &meshDataplane{
-		kubeClient: client.Kube(),
-		netServer:  fs,
-	}
+
+	server := getFakeDP(fs, client.Kube())
 
 	handlers := setupHandlers(ctx, client, server, "istio-system")
 	client.RunAndWait(ctx.Done())
@@ -618,10 +598,8 @@ func TestGetActiveAmbientPodSnapshotSkipsTerminatedJobPods(t *testing.T) {
 	client := kube.NewFakeClient(ns, enrolledNotRedirected, enrolledButTerminated)
 	fs := &fakeServer{}
 	fs.Start(ctx)
-	server := &meshDataplane{
-		kubeClient: client.Kube(),
-		netServer:  fs,
-	}
+
+	server := getFakeDP(fs, client.Kube())
 
 	handlers := setupHandlers(ctx, client, server, "istio-system")
 	client.RunAndWait(ctx.Done())
@@ -661,10 +639,8 @@ func TestAmbientEnabledReturnsPodIfEnabled(t *testing.T) {
 	client := kube.NewFakeClient(ns, pod)
 	fs := &fakeServer{}
 	fs.Start(ctx)
-	server := &meshDataplane{
-		kubeClient: client.Kube(),
-		netServer:  fs,
-	}
+
+	server := getFakeDP(fs, client.Kube())
 
 	handlers := setupHandlers(ctx, client, server, "istio-system")
 	client.RunAndWait(ctx.Done())
@@ -704,10 +680,8 @@ func TestAmbientEnabledReturnsNoPodIfNotEnabled(t *testing.T) {
 	client := kube.NewFakeClient(ns, pod)
 	fs := &fakeServer{}
 	fs.Start(ctx)
-	server := &meshDataplane{
-		kubeClient: client.Kube(),
-		netServer:  fs,
-	}
+
+	server := getFakeDP(fs, client.Kube())
 
 	handlers := setupHandlers(ctx, client, server, "istio-system")
 	client.RunAndWait(ctx.Done())
@@ -748,10 +722,8 @@ func TestAmbientEnabledReturnsErrorIfBogusNS(t *testing.T) {
 	client := kube.NewFakeClient(ns, pod)
 	fs := &fakeServer{}
 	fs.Start(ctx)
-	server := &meshDataplane{
-		kubeClient: client.Kube(),
-		netServer:  fs,
-	}
+
+	server := getFakeDP(fs, client.Kube())
 
 	handlers := setupHandlers(ctx, client, server, "istio-system")
 	client.RunAndWait(ctx.Done())
@@ -802,10 +774,7 @@ func TestExistingPodAddedWhenItPreExists(t *testing.T) {
 		"",
 	).Return(nil)
 
-	server := &meshDataplane{
-		kubeClient: client.Kube(),
-		netServer:  fs,
-	}
+	server := getFakeDP(fs, client.Kube())
 
 	handlers := setupHandlers(ctx, client, server, "istio-system")
 	client.RunAndWait(ctx.Done())
