+++ cni/pkg/nodeagent/server_test.go
@@ -23,11 +23,14 @@ import (
 	"time"
 
 	"github.com/stretchr/testify/mock"
+	"golang.org/x/sys/unix"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/kubernetes/fake"
 
+	"istio.io/istio/cni/pkg/ipset"
 	"istio.io/istio/pkg/config/constants"
 	"istio.io/istio/pkg/test/util/assert"
 )
@@ -56,13 +59,17 @@ func TestMeshDataplaneAddsAnnotationOnAdd(t *testing.T) {
 	).Return(nil)
 
 	server.Start(fakeCtx)
-	m := meshDataplane{
-		kubeClient: fakeClientSet,
-		netServer:  server,
-	}
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+
+	m := getFakeDPWithIPSet(server, fakeClientSet, set)
+	expectPodAddedToIPSet(fakeIPSetDeps, podIP, pod.ObjectMeta)
 
 	err := m.AddPodToMesh(fakeCtx, pod, podIPs, "")
 	assert.NoError(t, err)
+
+	fakeIPSetDeps.AssertExpectations(t)
 
 	pod, err = fakeClientSet.CoreV1().Pods("test").Get(fakeCtx, "test", metav1.GetOptions{})
 	assert.NoError(t, err)
@@ -93,13 +100,17 @@ func TestMeshDataplaneAddsAnnotationOnAddWithPartialError(t *testing.T) {
 
 	server.Start(fakeCtx)
 	fakeClientSet := fake.NewSimpleClientset(pod)
-	m := meshDataplane{
-		kubeClient: fakeClientSet,
-		netServer:  server,
-	}
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+
+	m := getFakeDPWithIPSet(server, fakeClientSet, set)
 
 	err := m.AddPodToMesh(fakeCtx, pod, podIPs, "")
 	assert.Error(t, err)
+
+	// as this is a partial add error we should NOT have added to the ipset
+	fakeIPSetDeps.AssertExpectations(t)
 
 	pod, err = fakeClientSet.CoreV1().Pods("test").Get(fakeCtx, "test", metav1.GetOptions{})
 	assert.NoError(t, err)
@@ -130,13 +141,17 @@ func TestMeshDataplaneDoesntAnnotateOnAddWithRealError(t *testing.T) {
 
 	server.Start(fakeCtx)
 	fakeClientSet := fake.NewSimpleClientset(pod)
-	m := meshDataplane{
-		kubeClient: fakeClientSet,
-		netServer:  server,
-	}
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+
+	m := getFakeDPWithIPSet(server, fakeClientSet, set)
 
 	err := m.AddPodToMesh(fakeCtx, pod, podIPs, "")
 	assert.Error(t, err)
+
+	// as this is a partial add error we should NOT have added to the ipset
+	fakeIPSetDeps.AssertExpectations(t)
 
 	pod, err = fakeClientSet.CoreV1().Pods("test").Get(fakeCtx, "test", metav1.GetOptions{})
 	assert.NoError(t, err)
@@ -157,13 +172,17 @@ func TestMeshDataplaneRemovePodRemovesAnnotation(t *testing.T) {
 	).Return(nil)
 
 	fakeClientSet := fake.NewSimpleClientset(pod)
-	m := meshDataplane{
-		kubeClient: fakeClientSet,
-		netServer:  server,
-	}
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+
+	m := getFakeDPWithIPSet(server, fakeClientSet, set)
+	expectPodRemovedFromIPSet(fakeIPSetDeps, pod.Status.PodIPs)
 
 	err := m.RemovePodFromMesh(fakeCtx, pod, false)
 	assert.NoError(t, err)
+
+	fakeIPSetDeps.AssertExpectations(t)
 
 	pod, err = fakeClientSet.CoreV1().Pods("test").Get(fakeCtx, "test", metav1.GetOptions{})
 	assert.NoError(t, err)
@@ -183,13 +202,17 @@ func TestMeshDataplaneRemovePodErrorDoesntRemoveAnnotation(t *testing.T) {
 	).Return(errors.New("fake error"))
 
 	fakeClientSet := fake.NewSimpleClientset(pod)
-	m := meshDataplane{
-		kubeClient: fakeClientSet,
-		netServer:  server,
-	}
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+
+	m := getFakeDPWithIPSet(server, fakeClientSet, set)
+	expectPodRemovedFromIPSet(fakeIPSetDeps, pod.Status.PodIPs)
 
 	err := m.RemovePodFromMesh(fakeCtx, pod, false)
 	assert.Error(t, err)
+
+	fakeIPSetDeps.AssertExpectations(t)
 
 	pod, err = fakeClientSet.CoreV1().Pods("test").Get(fakeCtx, "test", metav1.GetOptions{})
 	assert.NoError(t, err)
@@ -210,13 +233,17 @@ func TestMeshDataplaneDelPod(t *testing.T) {
 	).Return(nil)
 
 	fakeClientSet := fake.NewSimpleClientset()
-	m := meshDataplane{
-		kubeClient: fakeClientSet,
-		netServer:  server,
-	}
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+	m := getFakeDPWithIPSet(server, fakeClientSet, set)
+	expectPodRemovedFromIPSet(fakeIPSetDeps, pod.Status.PodIPs)
 
 	// pod is not in fake client, so if this will try to remove annotation, it will fail.
 	err := m.RemovePodFromMesh(fakeCtx, pod, true)
+
+	fakeIPSetDeps.AssertExpectations(t)
+
 	assert.NoError(t, err)
 }
 
@@ -234,14 +261,295 @@ func TestMeshDataplaneDelPodErrorDoesntPatchPod(t *testing.T) {
 	).Return(errors.New("fake error"))
 
 	fakeClientSet := fake.NewSimpleClientset()
-	m := meshDataplane{
-		kubeClient: fakeClientSet,
-		netServer:  server,
-	}
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+
+	m := getFakeDPWithIPSet(server, fakeClientSet, set)
+	expectPodRemovedFromIPSet(fakeIPSetDeps, pod.Status.PodIPs)
 
 	// pod is not in fake client, so if this will try to remove annotation, it will fail.
 	err := m.RemovePodFromMesh(fakeCtx, pod, true)
+
+	fakeIPSetDeps.AssertExpectations(t)
+
 	assert.Error(t, err)
+}
+
+func TestMeshDataplaneAddPodToHostNSIPSets(t *testing.T) {
+	pod := buildConvincingPod()
+
+	fakeCtx := context.Background()
+	server := &fakeServer{}
+	server.Start(fakeCtx)
+	fakeClientSet := fake.NewSimpleClientset()
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+	m := getFakeDPWithIPSet(server, fakeClientSet, set)
+
+	var podUID string = string(pod.ObjectMeta.UID)
+	ipProto := uint8(unix.IPPROTO_TCP)
+
+	fakeIPSetDeps.On("addIP",
+		"foo",
+		netip.MustParseAddr("99.9.9.9"),
+		ipProto,
+		podUID,
+		false,
+	).Return(nil)
+
+	fakeIPSetDeps.On("addIP",
+		"foo",
+		netip.MustParseAddr("2.2.2.2"),
+		ipProto,
+		podUID,
+		false,
+	).Return(nil)
+
+	podIPs := []netip.Addr{netip.MustParseAddr("99.9.9.9"), netip.MustParseAddr("2.2.2.2")}
+	_, err := m.addPodToHostNSIpset(pod, podIPs)
+	assert.NoError(t, err)
+
+	fakeIPSetDeps.AssertExpectations(t)
+}
+
+func TestMeshDataplaneAddPodIPToHostNSIPSetsReturnsErrorIfOneFails(t *testing.T) {
+	pod := buildConvincingPod()
+
+	fakeCtx := context.Background()
+	server := &fakeServer{}
+	server.Start(fakeCtx)
+	fakeClientSet := fake.NewSimpleClientset()
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+	m := getFakeDPWithIPSet(server, fakeClientSet, set)
+
+	var podUID string = string(pod.ObjectMeta.UID)
+	ipProto := uint8(unix.IPPROTO_TCP)
+
+	fakeIPSetDeps.On("addIP",
+		"foo",
+		netip.MustParseAddr("99.9.9.9"),
+		ipProto,
+		podUID,
+		false,
+	).Return(nil)
+
+	fakeIPSetDeps.On("addIP",
+		"foo",
+		netip.MustParseAddr("2.2.2.2"),
+		ipProto,
+		podUID,
+		false,
+	).Return(errors.New("bwoah"))
+
+	podIPs := []netip.Addr{netip.MustParseAddr("99.9.9.9"), netip.MustParseAddr("2.2.2.2")}
+	addedPIPs, err := m.addPodToHostNSIpset(pod, podIPs)
+	assert.Error(t, err)
+	assert.Equal(t, 1, len(addedPIPs), "only expected one IP to be added")
+
+	fakeIPSetDeps.AssertExpectations(t)
+}
+
+func TestMeshDataplaneRemovePodIPFromHostNSIPSets(t *testing.T) {
+	pod := buildConvincingPod()
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+
+	fakeIPSetDeps.On("clearEntriesWithIP",
+		"foo",
+		netip.MustParseAddr("3.3.3.3"),
+	).Return(nil)
+
+	fakeIPSetDeps.On("clearEntriesWithIP",
+		"foo",
+		netip.MustParseAddr("2.2.2.2"),
+	).Return(nil)
+
+	err := removePodFromHostNSIpset(pod, &set)
+	assert.NoError(t, err)
+	fakeIPSetDeps.AssertExpectations(t)
+}
+
+func TestMeshDataplaneSyncHostIPSetsPrunesNothingIfNoExtras(t *testing.T) {
+	pod := buildConvincingPod()
+
+	fakeCtx := context.Background()
+	server := &fakeServer{}
+	server.Start(fakeCtx)
+	fakeClientSet := fake.NewSimpleClientset()
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+	m := getFakeDPWithIPSet(server, fakeClientSet, set)
+
+	var podUID string = string(pod.ObjectMeta.UID)
+	ipProto := uint8(unix.IPPROTO_TCP)
+
+	// expectations
+	fakeIPSetDeps.On("addIP",
+		"foo",
+		netip.MustParseAddr("3.3.3.3"),
+		ipProto,
+		podUID,
+		false,
+	).Return(nil)
+
+	fakeIPSetDeps.On("addIP",
+		"foo",
+		netip.MustParseAddr("2.2.2.2"),
+		ipProto,
+		podUID,
+		false,
+	).Return(nil)
+
+	fakeIPSetDeps.On("listEntriesByIP",
+		"foo",
+	).Return([]netip.Addr{}, nil)
+
+	err := m.syncHostIPSets([]*corev1.Pod{pod})
+	assert.NoError(t, err)
+	fakeIPSetDeps.AssertExpectations(t)
+}
+
+func TestMeshDataplaneSyncHostIPSetsIgnoresPodIPAddErrorAndContinues(t *testing.T) {
+	pod1 := buildConvincingPod()
+	pod2 := buildConvincingPod()
+
+	pod2.ObjectMeta.SetUID("4455")
+
+	fakeClientSet := fake.NewSimpleClientset()
+
+	fakeCtx := context.Background()
+	server := &fakeServer{}
+	server.Start(fakeCtx)
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+	m := getFakeDPWithIPSet(server, fakeClientSet, set)
+
+	var pod1UID string = string(pod1.ObjectMeta.UID)
+	var pod2UID string = string(pod2.ObjectMeta.UID)
+	ipProto := uint8(unix.IPPROTO_TCP)
+
+	// First IP of first pod should error, but we should add the rest
+	fakeIPSetDeps.On("addIP",
+		"foo",
+		netip.MustParseAddr("3.3.3.3"),
+		ipProto,
+		pod1UID,
+		false,
+	).Return(errors.New("CANNOT ADD"))
+
+	fakeIPSetDeps.On("addIP",
+		"foo",
+		netip.MustParseAddr("2.2.2.2"),
+		ipProto,
+		pod1UID,
+		false,
+	).Return(nil)
+
+	fakeIPSetDeps.On("addIP",
+		"foo",
+		netip.MustParseAddr("3.3.3.3"),
+		ipProto,
+		pod2UID,
+		false,
+	).Return(errors.New("CANNOT ADD"))
+
+	fakeIPSetDeps.On("addIP",
+		"foo",
+		netip.MustParseAddr("2.2.2.2"),
+		ipProto,
+		pod2UID,
+		false,
+	).Return(nil)
+
+	fakeIPSetDeps.On("listEntriesByIP",
+		"foo",
+	).Return([]netip.Addr{}, nil)
+
+	err := m.syncHostIPSets([]*corev1.Pod{pod1, pod2})
+	assert.NoError(t, err)
+	fakeIPSetDeps.AssertExpectations(t)
+}
+
+func TestMeshDataplaneSyncHostIPSetsAddsNothingIfPodHasNoIPs(t *testing.T) {
+	pod := buildConvincingPod()
+
+	pod.Status.PodIP = ""
+	pod.Status.PodIPs = []corev1.PodIP{}
+
+	fakeCtx := context.Background()
+	server := &fakeServer{}
+	server.Start(fakeCtx)
+	fakeClientSet := fake.NewSimpleClientset()
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+	m := getFakeDPWithIPSet(server, fakeClientSet, set)
+
+	fakeIPSetDeps.On("listEntriesByIP",
+		"foo",
+	).Return([]netip.Addr{}, nil)
+
+	err := m.syncHostIPSets([]*corev1.Pod{pod})
+	assert.NoError(t, err)
+	fakeIPSetDeps.AssertExpectations(t)
+}
+
+func TestMeshDataplaneSyncHostIPSetsPrunesIfExtras(t *testing.T) {
+	pod := buildConvincingPod()
+
+	var podUID string = string(pod.ObjectMeta.UID)
+	ipProto := uint8(unix.IPPROTO_TCP)
+
+	fakeCtx := context.Background()
+	server := &fakeServer{}
+	server.Start(fakeCtx)
+	fakeClientSet := fake.NewSimpleClientset()
+
+	fakeIPSetDeps := ipset.FakeNLDeps()
+	set := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+	m := getFakeDPWithIPSet(server, fakeClientSet, set)
+
+	// expectations
+	fakeIPSetDeps.On("addIP",
+		"foo",
+		netip.MustParseAddr("3.3.3.3"),
+		ipProto,
+		podUID,
+		false,
+	).Return(nil)
+
+	fakeIPSetDeps.On("addIP",
+		"foo",
+		netip.MustParseAddr("2.2.2.2"),
+		ipProto,
+		podUID,
+		false,
+	).Return(nil)
+
+	// List should return one IP not in our "pod snapshot", which means we prune
+	fakeIPSetDeps.On("listEntriesByIP",
+		"foo",
+	).Return([]netip.Addr{
+		netip.MustParseAddr("2.2.2.2"),
+		netip.MustParseAddr("6.6.6.6"),
+		netip.MustParseAddr("3.3.3.3"),
+	}, nil)
+
+	fakeIPSetDeps.On("clearEntriesWithIP",
+		"foo",
+		netip.MustParseAddr("6.6.6.6"),
+	).Return(nil)
+
+	err := m.syncHostIPSets([]*corev1.Pod{pod})
+	assert.NoError(t, err)
+	fakeIPSetDeps.AssertExpectations(t)
 }
 
 func podWithAnnotation() *corev1.Pod {
@@ -338,4 +646,48 @@ func (wg *WaitGroup) Done() {
 
 func (wg *WaitGroup) C() <-chan struct{} {
 	return wg.done
+}
+
+func expectPodAddedToIPSet(ipsetDeps *ipset.MockedIpsetDeps, podIP netip.Addr, podMeta metav1.ObjectMeta) {
+	ipsetDeps.On("addIP",
+		"foo",
+		podIP,
+		uint8(unix.IPPROTO_TCP),
+		string(podMeta.UID),
+		false,
+	).Return(nil)
+}
+
+func expectPodRemovedFromIPSet(ipsetDeps *ipset.MockedIpsetDeps, podIPs []corev1.PodIP) {
+	for _, ip := range podIPs {
+		ipsetDeps.On("clearEntriesWithIP",
+			"foo",
+			netip.MustParseAddr(ip.IP),
+		).Return(nil)
+	}
+}
+
+func getFakeDPWithIPSet(fs *fakeServer, fakeClient kubernetes.Interface, fakeSet ipset.IPSet) *meshDataplane {
+	return &meshDataplane{
+		kubeClient:         fakeClient,
+		netServer:          fs,
+		hostsideProbeIPSet: fakeSet,
+	}
+}
+
+func getFakeDP(fs *fakeServer, fakeClient kubernetes.Interface) *meshDataplane {
+	fakeIPSetDeps := ipset.FakeNLDeps()
+
+	fakeIPSetDeps.On("addIP",
+		mock.Anything,
+		mock.Anything,
+		mock.Anything,
+		mock.Anything,
+		mock.Anything,
+	).Return(nil).Maybe()
+
+	fakeIPSetDeps.On("clearEntriesWithIP", mock.Anything, mock.Anything).Return(nil).Maybe()
+	fakeSet := ipset.IPSet{Name: "foo", Deps: fakeIPSetDeps}
+
+	return getFakeDPWithIPSet(fs, fakeClient, fakeSet)
 }
